# api/openapi.yaml
openapi: 3.0.3
info:
  title: Live Streaming Tipping Platform API
  description: |
    Secure, scalable live-streaming platform where performers broadcast video, 
    viewers watch, and viewers tip virtual tokens in real time.
    
    ## Key Features
    - Real-time WebRTC streaming with SFU
    - Virtual token economy with instant tipping
    - Age verification and KYC compliance
    - Automated and human moderation
    - PCI-compliant payment processing
    
    ## Authentication
    All endpoints require JWT authentication except registration, login, and public endpoints.
    Include the JWT token in the Authorization header: `Bearer <token>`
    
    ## Rate Limiting
    - General API: 1000 requests/hour per user
    - Streaming endpoints: 100 requests/minute per user
    - Payment endpoints: 10 requests/minute per user
    - Tipping: 60 tips/minute per user
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@livepanty.com
  license:
    name: Proprietary
servers:
  - url: https://api.livepanty.com/v1
    description: Production server
  - url: https://staging-api.livepanty.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user account
      description: Create a new user account with email verification
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and return JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user profile
      description: Retrieve current authenticated user's profile
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /admin/users:
    get:
      tags:
        - Admin
      summary: Get all users (Admin only)
      description: Retrieve list of all users with pagination
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [viewer, performer, admin, moderator]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, suspended, banned, pending_verification]
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Users retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserSummary'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
...
  /admin/analytics/overview:
    get:
      tags:
        - Admin
      summary: Get platform analytics (Admin only)
      description: Retrieve platform-wide analytics and metrics
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [day, week, month, year]
            default: week
      responses:
        '200':
          description: Analytics retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlatformAnalytics'

# ============================================================================
# COMPONENTS
# ============================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - display_name
        - date_of_birth
        - country
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          example: "SecurePass123!"
        display_name:
          type: string
          minLength: 2
          maxLength: 100
          example: "John Doe"
        username:
          type: string
          minLength: 3
          maxLength: 50
          pattern: "^[a-zA-Z0-9_]+$"
          example: "johndoe"
        date_of_birth:
          type: string
          format: date
          example: "1990-01-01"
        country:
          type: string
          pattern: "^[A-Z]{2}$"
          example: "US"
        role:
          type: string
          enum: [viewer, performer]
          default: viewer
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          example: "SecurePass123!"
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        expires_in:
          type: integer
          example: 3600
        token_type:
          type: string
          example: "Bearer"
        user:
          $ref: '#/components/schemas/UserProfile'

    # ============================================================================
    # USER SCHEMAS
    # ============================================================================
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [viewer, performer, admin, moderator]
        status:
          type: string
          enum: [active, suspended, banned, pending_verification]
        country:
          type: string
        timezone:
          type: string
        date_of_birth:
          type: string
          format: date
        profile_image_url:
          type: string
          format: uri
        bio:
          type: string
        is_public:
          type: boolean
        email_verified:
          type: boolean
        two_factor_enabled:
          type: boolean
        last_login_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
        username:
          type: string
        role:
          type: string
        status:
          type: string
        country:
          type: string
        email_verified:
          type: boolean
        last_login_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        token_balance:
          type: integer
        kyc_status:
          type: string

    # ============================================================================
    # KYC SCHEMAS
    # ============================================================================
...
    PlatformAnalytics:
      type: object
      properties:
        period:
          type: string
        total_users:
          type: integer
        active_performers:
          type: integer
        total_streams:
          type: integer
        active_streams:
          type: integer
        total_viewers:
          type: integer
        total_tips:
          type: integer
        total_tokens_tipped:
          type: integer
        total_revenue:
          type: number
          format: float
        average_stream_duration:
          type: integer
        kyc_pending:
          type: integer
        moderation_queue:
          type: integer
        top_categories:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              count:
                type: integer
              revenue:
                type: number
                format: float

    # ============================================================================
    # ERROR SCHEMAS
    # ============================================================================
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time